pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-gene-enrichment"
version = "0.1.0"
description = "A comprehensive tool for running gene enrichment analysis workflows using multiple enrichment tools and AI-powered summarization"
readme = "README.md"
license = {text = "CC-BY-NC-4.0"}
authors = [
    {name = "Calvin Boyce", email = "calvin.boyce@harvard.edu"}
]
maintainers = [
    {name = "Calvin Boyce", email = "calvin.boyce@harvard.edu"}
]
keywords = ["bioinformatics", "gene-enrichment", "ai", "genomics", "enrichment-analysis"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering :: Bio-Informatics",
    "License :: Other/Proprietary License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Operating System :: OS Independent",
]
requires-python = ">=3.8"
dependencies = [
    "openai>=1.0.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "biopython>=1.82",
    "gprofiler-official>=1.0.0",
    "openpyxl>=3.1.5",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/calvinboyce/ai-gene-enrichment"
Repository = "https://github.com/calvinboyce/ai-gene-enrichment"
Documentation = "https://github.com/calvinboyce/ai-gene-enrichment#readme"
Issues = "https://github.com/calvinboyce/ai-gene-enrichment/issues"



[tool.setuptools.packages.find]
where = ["."]
include = ["aige*"]

[tool.setuptools.package-data]
aige = ["*.txt", "*.md"]

[tool.black]
line-length = 88
target-version = ['py38']

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
